feat: a token count nobody reported is not a token count of zero
Usage carried four primitive counts, so "the provider does not report caching"
and "the cache was not used" were the same number. They are not the same fact,
and the difference is the one a graph turns on: invented zeros averaged into a
cache-hit rate read as a cache working badly rather than a provider that keeps
no such books. Every count is now a nullable Integer, and null means the
provider did not say.

Integer rather than Long because a single call reporting two billion tokens is
not a thing we need to plan for.

The lie was already in the code in four places, each now saying what it means
with Usage.unreported(): ModelReplies when a stream ends without a usage event,
and the per-stream accumulators in the OpenAI, Gemini and Bedrock adapters,
which opened claiming a call had so far cost zero rather than that nothing had
been reported yet. A turn whose model never reported now ends unreported too,
instead of summing to zero.

Per-vendor truth stays per-vendor, which is what these adapters are for.
Anthropic and Bedrock report their cache pair and support caching, so an
omitted field there means nothing was cached: those stay zero, and their sums
still have numbers to add. The OpenAI wire is the other case — LM Studio sends
no prompt_tokens_details at all — so there the counts stay unreported.

Also fixes a real gap this uncovered: the OpenAI adapter dropped cached tokens
entirely. Its comment still read "Usage carries two numbers now", left behind
when Usage was narrowed and never revisited when it was restored, so
prompt_tokens_details.cached_tokens never landed. A test named for that path
had been quietly asserting it did not happen.

Observed now records nothing for a count nobody reported — no span attribute,
no histogram point. A REPORTED zero is still written, for the reason it always
was: a missing series and a genuine zero look identical on a graph.

Verified against a live LM Studio: input 25, output 63, cache read and write
both null rather than zero.

+++ nessy-model-anthropic/src/main/java/org/jwcarman/nessy/model/anthropic/AnthropicStream.java
@@ -79,9 +79,9 @@ public final class AnthropicStream implements ModelStream {
     private final Iterator<RawMessageStreamEvent> events;
     private final Deque<ModelEvent> pending = new ArrayDeque<>();
     private final Map<Long, PendingToolUse> toolUsesByIndex = new HashMap<>();
-    private long inputTokens;
-    private long cacheReadInputTokens;
-    private long cacheWriteInputTokens;
+    private int inputTokens;
+    private int cacheReadInputTokens;
+    private int cacheWriteInputTokens;
     private boolean turnEnded;
 
     private TranslatingIterator(Iterator<RawMessageStreamEvent> events) {
@@ -139,10 +139,13 @@ public final class AnthropicStream implements ModelStream {
     private void translate(RawMessageStreamEvent event) {
       if (event.isMessageStart()) {
         var usage = event.asMessageStart().message().usage();
+        // Absent cache counts are ZERO here rather than null, unlike the OpenAI wire: Anthropic
+        // supports prompt caching and reports these fields, so their absence is the vendor saying
+        // nothing was cached — and the sum below needs a number anyway.
         readUsage(
-            usage.inputTokens(),
-            usage.cacheReadInputTokens().orElse(0L),
-            usage.cacheCreationInputTokens().orElse(0L));
+            Math.toIntExact(usage.inputTokens()),
+            Math.toIntExact(usage.cacheReadInputTokens().orElse(0L)),
+            Math.toIntExact(usage.cacheCreationInputTokens().orElse(0L)));
       } else if (event.isContentBlockStart()) {
         translateContentBlockStart(event.asContentBlockStart());
       } else if (event.isContentBlockDelta()) {
@@ -180,7 +183,7 @@ public final class AnthropicStream implements ModelStream {
      * @param cacheRead the vendor's {@code cache_read_input_tokens}
      * @param cacheWrite the vendor's {@code cache_creation_input_tokens}
      */
-    private void readUsage(long uncachedInputTokens, long cacheRead, long cacheWrite) {
+    private void readUsage(int uncachedInputTokens, int cacheRead, int cacheWrite) {
       // Anthropic's own input_tokens EXCLUDES both cache counts, so the whole is the sum.
       inputTokens = uncachedInputTokens + cacheRead + cacheWrite;
       cacheReadInputTokens = cacheRead;
@@ -253,7 +256,7 @@ public final class AnthropicStream implements ModelStream {
       var usage =
           new Usage(
               inputTokens,
-              event.usage().outputTokens(),
+              Math.toIntExact(event.usage().outputTokens()),
               cacheReadInputTokens,
               cacheWriteInputTokens);
       // A refusal is its own event now, not a stop reason: StopReason names only the three ways a
